+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/combat/fletching/entity/projectile/TorchArrow.java
@@ -31,7 +31,7 @@ public class TorchArrow extends Arrow {
 
     @Override
     protected void onHitBlock(BlockHitResult pResult) {
-        if (this.level().isClientSide) {
+        if (this.level().isClientSide || !Fletching.torchArrowsPlaceTorches) {
             super.onHitBlock(pResult);
             return;
         }